suppression/téléchargement de versions

+++ www/workspace/file_versions.js
@@ -1,35 +1,36 @@
 angular.module('ent.workspace_file_versions',['ent.workspace_service'])
 
-.controller('FileVersionCtrl', function($scope, $rootScope, WorkspaceService,$ionicLoading){
+.controller('FileVersionCtrl', function($scope, $rootScope, WorkspaceService,$ionicLoading, domainENT){
 
   getData($rootScope.doc._id)
 
   $scope.downloadVersion = function(version){
-    $scope.getConfirmPopup('Télécharger', 'Voulez-vous télécharger cette version de ce fichier ?').then(function(res){
+    $scope.getConfirmPopup('Télécharger', 'Voulez-vous télécharger cette version de ce fichier ?', $rootScope.translationWorkspace["cancel"], 'OK' ).then(function(res){
       console.log(res);
       if(res){
-
-          // $scope.downloadAttachment = function (id){
-          //   var attachmentUrl = domainENT+"/conversation/message/"+$scope.email.id+"/attachment/"+id;
-          //   var attachment = findElementById($scope.email.attachments, id);
-          //   $scope.downloadFile(attachment.filename, attachmentUrl,attachment.contentType);
-          // }
+        var versionUrl = domainENT+"/workspace/document/"+$rootScope.doc._id+"/revision/"+version._id;
+        $scope.downloadFile(version.filename, versionUrl,version.metadata["content-type"]);
       }
     })
   }
 
   $scope.deleteVersion = function(idVersion){
-    $ionicLoading.show({
-      template: '<ion-spinner icon="android"/>'
-    });
-    WorkspaceService.deleteVersion($rootScope.doc._id, idVersion).then(function(result){
-      console.log(result.data);
-      getData($rootScope.doc._id)
-      $ionicLoading.hide()
-    }, function(err){
-      $ionicLoading.hide()
-      $scope.showAlertError()
-    });
+    $scope.getConfirmPopup($rootScope.translationWorkspace["workspace.delete.version"], 'Êtes-vous sûr(e) de vouloir supprimer cette version de ce fichier ?', $rootScope.translationWorkspace["cancel"], 'OK' ).then(function(res){
+      console.log(res);
+      if(res){
+        $ionicLoading.show({
+          template: '<ion-spinner icon="android"/>'
+        });
+        WorkspaceService.deleteVersion($rootScope.doc._id, idVersion).then(function(result){
+          console.log(result.data);
+          getData($rootScope.doc._id)
+          $ionicLoading.hide()
+        }, function(err){
+          $ionicLoading.hide()
+          $scope.showAlertError()
+        });
+      }
+    })
   }
 
   function getData(fileId){
